moved logic to Scene to access Threlte ctx and cache
