+++ site/src/components/PaperStack.svelte
@@ -1,46 +1,15 @@
 <script lang="ts">
-	import { createPapersFromItems, replaceBottomPaper } from '$lib/initPapers'
 	import Scene from '$lib/Scene/Scene.svelte'
 	import type { ArchiveItem } from '$lib/types'
-	import { Canvas, type ThrelteContext } from '@threlte/core'
-	import { random } from 'lodash-es'
-	import { onDestroy, onMount } from 'svelte'
+	import { Canvas } from '@threlte/core'
 	import { NoToneMapping, SRGBColorSpace } from 'three'
-	import { lastActivity, papers } from '../store'
+	import { papers } from '../store'
 
 	export let items: ArchiveItem[]
-
-	let three: ThrelteContext
-
-	let timeout: NodeJS.Timeout
-
-	onMount(() => {
-		createPapersFromItems(three, items)
-	})
-
-	const createPaperTimeout = () => {
-		timeout = setTimeout(() => {
-			replaceBottomPaper(three)
-			createPaperTimeout()
-		}, random(5000, 8000))
-	}
-
-	onDestroy(
-		lastActivity.subscribe(() => {
-			if (timeout) clearTimeout(timeout)
-			createPaperTimeout()
-		})
-	)
-
-	const cancelTimeout = () => {
-		if (timeout) clearTimeout(timeout)
-	}
 </script>
 
-<svelte:window on:visibilitychange={cancelTimeout} />
-
-<Canvas shadows={true} bind:ctx={three} colorSpace={SRGBColorSpace} toneMapping={NoToneMapping}>
-	<Scene />
+<Canvas shadows={true} colorSpace={SRGBColorSpace} toneMapping={NoToneMapping}>
+	<Scene {items} />
 </Canvas>
 
 <div class="debug" class:visible={import.meta.env.DEV}>
